Add poisson noise to simulate CT noise distribution
CT noise is best modeled by a Poisson distribution. According to
Wikipedia:

> "The Poisson distribution is popular for modelling the number of times
> an event occurs in an interval of time or space."

In the case of a CT or x-ray, this corresponds to the number of times a
photon will strike the sensor during the interval of imaging.

Required the definition of two parameters:

`lambda`: An approximation to the number of photons that are passed
through each sensor. This is the "expected interval" of the Poisson
distribution.

`contrast`: This is the relative intensity of the fiducial when compared
to the noise.